binary-classification/frameworks/autosklearn/requirements.txt
setuptools
pytest>=4.6
Cython

numpy>=1.9.0
scipy>=0.14.1

scikit-learn>=0.22.0,<0.23

lockfile
joblib
psutil
pyyaml
liac-arff
pandas<1.0

ConfigSpace>=0.4.0,<0.5
pynisher>=0.4.2
pyrfr>=0.7,<0.9
smac>=0.12

category-encoders>=2.2.2
pylint>=2.4.4
xgboost>=1.0.2
tqdm>=4.43.0
auto-sklearn==0.8.0